component for versions
